requirements.txt
Flask
flask-mysql
pyyaml
